flask-server/requirements.txt
flask

# OCR
pytesseract
Pillow
PyMuPDF

# LangChain core + plugins
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters

# Embedding + vector DB
openai
faiss-cpu